Improve Makefile

+++ Makefile
@@ -1,7 +1,15 @@
+# Simple makefile to create a tarball of the msman helper scripts
+
 all:
-	cp -r .msman msman
-	tar -czf msman-hepler.tar.gz msman
-	rm -rf msman
+	source msman.sh
+	source .msman/version.sh
+	if [[ $CURRENT_SCRIPT_VERSION == $EXTRA_SCRIPTS_VERSION]]; then
+		cp -r .msman msman
+		tar -czf msman-hepler.tar.gz msman
+		rm -rf msman
+	else
+		echo "msman.sh and .msman/version.sh are not in sync"
+	fi
 
 clean:
 	rm -rf msman
